Update date formatting logic
