+++ Wire-iOS/Resources/Base.lproj/Localizable.strings
@@ -489,7 +489,10 @@
 "content.message.open" = "Open";
 "content.message.copy" = "Copy";
 "content.message.download" = "Download";
+
+// Reply
 "content.message.reply" = "Reply";
+"content.message.reply.original_timestamp" = "Original message from %@";
 
 "content.message.open_link_alert.title" = "Visit Link";
 "content.message.open_link_alert.message" = "This will take you to\n%@";
